Fix: CollectionView를 분리 후 Modal을 BillModalViewController로 파일 네이밍 변경
