+++ TodayBill/BillModal/BillModalCollectionViewController.swift
@@ -1,0 +1,95 @@
+//
+//  BillModalCollectionViewController.swift
+//  TodayBill
+//
+//  Created by 김건호 on 1/8/25.
+//
+
+import UIKit
+
+import UIKit
+
+class BillModalCollectionViewController: UIViewController {
+    // Properties
+    var dataRows = [Row]()
+    var filteredDataRows = [Row]()
+    var favoriteData = [Row]()
+    weak var delegate: BillModalViewControllerDelegate?
+
+    private lazy var collectionView: UICollectionView = {
+        let layout = UICollectionViewFlowLayout()
+        layout.sectionInset = UIEdgeInsets(top: 0, left: 20, bottom: 0, right: 20)
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+        collectionView.backgroundColor = .white
+        collectionView.register(UICollectionViewCell.self, forCellWithReuseIdentifier: "cell")
+        collectionView.delegate = self
+        collectionView.dataSource = self
+        return collectionView
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        setupSubviews()
+        setupConstraints()
+    }
+
+    private func setupSubviews() {
+        view.addSubview(collectionView)
+    }
+
+    private func setupConstraints() {
+        collectionView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            collectionView.topAnchor.constraint(equalTo: view.topAnchor),
+            collectionView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            collectionView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+            collectionView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
+        ])
+    }
+
+    // MARK: - Helper Methods
+    func reloadData() {
+        collectionView.reloadData()
+    }
+}
+
+extension BillModalCollectionViewController: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return filteredDataRows.count
+    }
+
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "cell", for: indexPath)
+        // 기존 하위 뷰 제거
+        for subview in cell.contentView.subviews {
+            subview.removeFromSuperview()
+        }
+        let label = UILabel()
+        label.text = filteredDataRows[indexPath.item].BILL_NAME
+        label.textAlignment = .center
+        label.translatesAutoresizingMaskIntoConstraints = false
+        cell.contentView.addSubview(label)
+        NSLayoutConstraint.activate([
+            label.topAnchor.constraint(equalTo: cell.contentView.topAnchor),
+            label.leadingAnchor.constraint(equalTo: cell.contentView.leadingAnchor),
+            label.trailingAnchor.constraint(equalTo: cell.contentView.trailingAnchor),
+            label.bottomAnchor.constraint(equalTo: cell.contentView.bottomAnchor)
+        ])
+        cell.layer.borderWidth = 1
+        cell.layer.borderColor = UIColor.gray.cgColor
+        cell.layer.cornerRadius = 10
+        cell.clipsToBounds = true
+        return cell
+    }
+
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        let selectedRow = filteredDataRows[indexPath.item]
+        delegate?.favoriteDataUpdated(favoriteData) // Delegate 호출
+    }
+
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let padding: CGFloat = 10
+        let cellWidth = collectionView.bounds.width - padding * 2
+        return CGSize(width: cellWidth, height: 50)
+    }
+}
